Completed 3.1 integration of Contrib
completed memoryindex (contrib now complete)

+++ src/core/util/LuceneThread.cpp
@@ -113,9 +113,9 @@ namespace Lucene
     int64_t LuceneThread::currentId()
     {
         #if defined(_WIN32) || defined(_WIN64)
-        return GetCurrentThreadId();
+        return (int64_t)GetCurrentThreadId();
         #else
-        return pthread_self();
+        return (int64_t)pthread_self();
         #endif
     }
